Refactor type of value

+++ db/migrate/20211124143633_create_locations.rb
@@ -1,8 +1,8 @@
 class CreateLocations < ActiveRecord::Migration[6.1]
   def change
     create_table :locations do |t|
-      t.integer :lat
-      t.integer :lon
+      t.float :lat
+      t.float :lon
       t.string :city
       t.string :state
       t.references :weather, null: false, foreign_key: true
